Fix newline issue (and remove database)
An extra newline is required at the end of each set in order for it to
paste correctly into (e.g.) Showdown. This does result in extra spacing,
I may or may not come up with a more elegant solution.

Also the database (pokepaste.db) should never have been synced.