Add highlighting for type-resist berries
Type-resist berries are now highlighted according to the type they
resist, e.g. a Yache Berry is light blue for ice resist.

This is particularly useful since the f***ing berry names are so hard to
remember.

I am yet to check whether this matches, or conflicts with, Natural Gift
typings for berries. Ah well ¯\_(ツ)_/¯