Clean Pokemon parsing
I added the function `parse_pokemon` which returns a dictionary object for any single Pokemon import. Working with an object is more convenient than working with a string.

I am not as familiar with Python as I am with Java/Javascript so styling conventions may be incorrect. Parsing for Hidden Power types and nature is incomplete. I do not feel confident in fully integrating the function in yet because I do not understand the original `format_paste` function.