Add support for SQLite3 and MySQL in parallel
To be fair, this does double the chances of something going wrong, but
it will also make migration (if and when that ever happens) a hell of a
lot easier.

Also, make a bunch of lines too long for PEP 8, and removed a piece of
ghost code that was doing literally nothing.