Refactor URL obfuscation into cipher.go
Hopefully this will allow a standalone URL {en,de}cryption tool, once I
figure out a clean way to namespace the part of the program I want to be
reusable. Expect some moving of things around while I work that out.

As a side-effect, standardises on uint64 everywhere for database IDs,
because that's what the actual database uses (BIGINT UNSIGNED).